Step 2 End update for pubsub DAPR component

+++ 02_Azure_Container_App/End/TPaperOrders/Models/DeliveryModel.cs
@@ -7,7 +7,7 @@ namespace TPaperOrders
     {
         public int Id { get; set; }
 
-        public double Number { get; set; }
+        public decimal Number { get; set; }
 
         public int ClientId { get; set; }
